feat: 코드 리뷰 수정사항 반영
