+++ woo-chang/java_racing_car/src/main/java/com/woochang/racingcar/domain/Car.java
@@ -1,4 +1,4 @@
-package com.woochang.racingcar;
+package com.woochang.racingcar.domain;
 
 public class Car {
