+++ woo-chang/java_racing_car/src/main/java/com/woochang/racingcar/domain/User.java
@@ -1,4 +1,4 @@
-package com.woochang.racingcar;
+package com.woochang.racingcar.domain;
 
 import java.util.List;
