+++ woo-chang/java_racing_car/src/main/java/com/woochang/racingcar/game/GameInput.java
@@ -1,4 +1,4 @@
-package com.woochang.racingcar;
+package com.woochang.racingcar.game;
 
 import java.util.Arrays;
 import java.util.List;
