+++ woo-chang/java_racing_car/src/test/java/com/woochang/racingcar/domain/CarTest.java
@@ -1,4 +1,4 @@
-package com.woochang.racingcar;
+package com.woochang.racingcar.domain;
 
 import static org.assertj.core.api.Assertions.*;
